feature:add interface with auth center. issue #1771
